requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
cffi==1.15.1
charset-normalizer==3.1.0
discord==2.2.2
discord.py==2.2.2
ffmpeg==1.4
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
yarl==1.8.2
